requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
mysqlclient>=2.2
django-filter>=24.2
openpyxl>=3.1
Pillow>=10.4
drf-spectacular>=0.27
python-dotenv>=1.0
akshare>=1.16
free-proxy>=1.1
